[refactor/#2]: 회원가입 페이지 아톰 수정
