--- app/components/atoms/CheckButton/CheckButton.style.ts
@@ -1,9 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  button: {
-    marginRight: 12,
-  },
-});
-
-export default styles;
